modify run.sh to support multiple podcast urls and transcode env
This update enhances the run.sh script to support multiple podcast URLs, which can be separated by spaces. This was supposed to be working before, but did not work for me.

Additionally, the TRANSCODE environment variable, as defined in the docker-compose file, is now correctly passed to the spodcast command within the script.